requirements.txt
flask
pandas
numpy
streamlit
plotly
nltk
markdown
scikit-learn
python-dotenv==1.0.1
google-generativeai==0.3.2
dotenv
